ml-predict-service/requirements.txt
nats-py==2.9.0
scikit-learn==1.4.0
torch==2.2.0
joblib==1.3.2
numpy==1.26.0